Allow cache lookups of name to UUID.
Previously only UUID -> name could be looked up, which meant the ProfileService would be queried every time a name needed to be looked up, even if that name was already cached.

Note that this currently means that old UUIDs may be returned for names that were changed and then taken by another player.