Create a service that checks Bukkit profiles first.
Unlike the previous changes to CacheForwardingService, this new class will force
the cached profiles to be updated when an online player is queried. If the player
is not online, it will fall back to cache or Mojang API look-ups as before.

Note that this isn't immune to the caveat mentioned in the last commit, but it
will at least reduce the frequency as long as players log on more often than
changed name holds expire. Of course, if the player doesn't log on in that
amount of time, they won't be using their WorldGuard region or whatever anyway.